+++ spec/models/project_spec.rb
@@ -102,6 +102,24 @@ RSpec.describe Project, type: :model do
           expect(subject.errors[:target_completion_date]).to include(I18n.t("activerecord.errors.models.project.attributes.target_completion_date.must_be_first_of_the_month"))
         end
       end
+
+      context "when date is today" do
+        subject { build(:project, target_completion_date: Date.today) }
+
+        it "is invalid" do
+          expect(subject).to_not be_valid
+          expect(subject.errors[:target_completion_date]).to include(I18n.t("activerecord.errors.models.project.attributes.target_completion_date.must_be_in_the_future"))
+        end
+      end
+
+      context "when date is in the past" do
+        subject { build(:project, target_completion_date: Date.yesterday) }
+
+        it "is invalid" do
+          expect(subject).to_not be_valid
+          expect(subject.errors[:target_completion_date]).to include(I18n.t("activerecord.errors.models.project.attributes.target_completion_date.must_be_in_the_future"))
+        end
+      end
     end
 
     describe "#team_leader" do